+++ src/components/scenes/ManageTokensScene.tsx
@@ -1,10 +1,11 @@
 import { EdgeCurrencyWallet } from 'edge-core-js'
 import * as React from 'react'
-import { View } from 'react-native'
+import { SectionList, View } from 'react-native'
 import { FlatList } from 'react-native-gesture-handler'
 
 import { PREFERRED_TOKENS, SPECIAL_CURRENCY_INFO } from '../../constants/WalletAndCurrencyConstants'
 import { useHandler } from '../../hooks/useHandler'
+import { useRowLayout } from '../../hooks/useRowLayout'
 import { useWalletName } from '../../hooks/useWalletName'
 import { useWatch } from '../../hooks/useWatch'
 import { lstrings } from '../../locales/strings'
@@ -21,11 +22,23 @@ import { FilledTextInput } from '../themed/FilledTextInput'
 import { ManageTokensRow } from '../themed/ManageTokensRow'
 import { SceneHeader } from '../themed/SceneHeader'
 import { Title } from '../themed/Title'
+import { WalletListSectionHeader } from '../themed/WalletListSectionHeader'
 import { ButtonsViewUi4 } from '../ui4/ButtonsViewUi4'
 import { CryptoIconUi4 } from '../ui4/CryptoIconUi4'
 
+/**
+ * walletId: ID of the wallet whose tokens we are managing
+ * newTokenIds: We routed from a notification that was triggered from detecting
+ * these new tokenIds having nonzero balance. Show these at the top of the list.
+ */
 export interface ManageTokensParams {
   walletId: string
+  newTokenIds?: string[]
+}
+
+interface Section {
+  title: string
+  data: string[]
 }
 
 interface Props extends EdgeSceneProps<'manageTokens'> {
@@ -33,7 +46,8 @@ interface Props extends EdgeSceneProps<'manageTokens'> {
 }
 
 function ManageTokensSceneComponent(props: Props) {
-  const { navigation, wallet } = props
+  const { navigation, route, wallet } = props
+  const { newTokenIds } = route.params
 
   const theme = useTheme()
   const styles = getStyles(theme)
@@ -97,6 +111,35 @@ function ManageTokensSceneComponent(props: Props) {
     })
   }, [allTokens, searchValue, sortedTokenIds])
 
+  // Split the list of tokens based on if there were auto-detected tokens given
+  const autoDetectedTokenIds = React.useMemo(
+    () => (newTokenIds ? filteredTokenIds.filter(filteredTokenId => newTokenIds.includes(filteredTokenId)) : []),
+    [filteredTokenIds, newTokenIds]
+  )
+
+  const extraData = React.useMemo(() => ({ allTokens, enabledTokenSet, customTokens }), [allTokens, enabledTokenSet, customTokens])
+
+  const sectionList = React.useMemo<Section[] | null>(() => {
+    if (autoDetectedTokenIds.length === 0) {
+      return null // Non-sectioned list off raw tokenIds
+    } else {
+      return [
+        {
+          title: lstrings.managetokens_detected_tokens_header,
+          data: autoDetectedTokenIds
+        },
+        {
+          title: lstrings.managetokens_all_tokens_header,
+
+          // Omit the auto-detected tokens we're already showing above
+          data: filteredTokenIds.filter(filteredTokenId => !autoDetectedTokenIds.includes(filteredTokenId))
+        }
+      ]
+    }
+  }, [autoDetectedTokenIds, filteredTokenIds])
+
+  const handleItemLayout = useRowLayout()
+
   // Goes to the add token scene:
   const handleAdd = useHandler(() => {
     navigation.navigate('editToken', {
@@ -116,12 +159,15 @@ function ManageTokensSceneComponent(props: Props) {
         isCustom={customTokens[tokenId] != null}
         isEnabled={enabledTokenSet.has(tokenId)}
         token={allTokens[tokenId]}
-        tokenId={item.item}
+        tokenId={tokenId}
       />
     )
   })
 
-  const extraData = React.useMemo(() => ({ allTokens, enabledTokenSet, customTokens }), [allTokens, enabledTokenSet, customTokens])
+  // Render a section header
+  const renderSectionHeader = useHandler((section: { section: Section }) => {
+    return <WalletListSectionHeader title={section.section.title} />
+  })
 
   return (
     <SceneWrapper>
@@ -137,7 +183,20 @@ function ManageTokensSceneComponent(props: Props) {
           onChangeText={setSearchValue}
         />
       </SceneHeader>
+      {sectionList == null ? (
         <FlatList data={filteredTokenIds} extraData={extraData} keyExtractor={keyExtractor} renderItem={renderRow} style={styles.list} />
+      ) : (
+        <SectionList
+          getItemLayout={handleItemLayout}
+          extraData={extraData}
+          keyboardDismissMode="on-drag"
+          keyboardShouldPersistTaps="handled"
+          renderItem={renderRow}
+          renderSectionHeader={renderSectionHeader}
+          sections={sectionList}
+          style={styles.sectionList}
+        />
+      )}
       {!isCustomTokensSupported ? null : (
         <>
           <DividerLine marginRem={[0, 1]} />
@@ -169,6 +228,10 @@ const getStyles = cacheStyles((theme: Theme) => ({
   subTitle: {
     color: theme.secondaryText,
     fontSize: theme.rem(0.85)
+  },
+  sectionList: {
+    marginTop: theme.rem(1),
+    marginHorizontal: theme.rem(0.5)
   }
 }))